+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Local hazard visualization [#143](https://github.com/IN-CORE/pyincore-viz/issues/143)
 
+### Changed
+- Clean dependencies [#145](https://github.com/IN-CORE/pyincore-viz/issues/145)
+
 ## [1.8.4] - 2023-11-08
 ### Updated
 - Basemap api updated to use open street map [#138](https://github.com/IN-CORE/pyincore-viz/issues/138)